requirements.txt
# Requirements completo para sistema RAG + Chatbot
# Instale com: pip install -r requirements.txt

# === CORE DEPENDENCIES ===
# Processamento de texto e chunking
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# OpenAI para embeddings e chat
openai==1.10.0

# Pinecone para banco vetorial (usar pinecone-client)
pinecone-client>=3.0.0

# Streamlit para interface web
streamlit==1.31.0

# Utilitários essenciais
python-dotenv==1.0.0
tiktoken==0.5.2

# === DEPENDENCIES ADICIONAIS ===
# Para processamento de texto
numpy==1.24.0
pandas==2.1.0

# Para requests e HTTP
requests==2.31.0
urllib3==2.0.7

# Para encoding e tokenização
charset-normalizer==3.3.0
certifi==2023.11.17

# Para JSON e serialização
pydantic==2.5.0
pydantic-core==2.14.0

# Para async operations
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3

# Para type hints e utilities
typing-extensions==4.8.0
attrs==23.1.0

# Para data manipulation
pytz==2023.3
six==1.16.0

# Para rich text e progress bars
rich==13.7.0
click==8.1.7

# Para caching e performance
cachetools==5.3.2

# === OPTIONAL BUT RECOMMENDED ===
# Para melhor logging
colorama==0.4.6

# Para handling de arquivos
pathlib2==2.3.7

# Para regex avançado
regex==2023.10.3